+++ sparkles-core/src/consts.rs
@@ -1,2 +1,2 @@
-// build.rs
+//! Auto-generated from build.rs
 include!(concat!(env!("OUT_DIR"), "/consts.rs"));