+++ CHANGELOG.md
@@ -4,14 +4,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [Unreleased]
-- [sparkles] Removed TCP support
+## [0.1.4] - 2024-09-28
 - [sparkles] Added file saving support
 - [sparkles] **[WIP]** Added UDP sender support with configuration 
 - [sparkles] Add sender config to SparklesConfig
-- [sparkles-core] **[WIP]** aarch64 timestamps support
+- [sparkles] Removed TCP support
+- [sparkles] Add `self-tracing` feature
+- [sparkles] Improve flushing performance: Largely reduce TLS operations blocking during global storage flushing.
+- [sparkles] Fix potential `ticks_per_sec` overflow
+- [sparkles-core] aarch64 timestamps support
+- [sparkles-core] Add `try_flush` and `is_buffer_available` to the global storage ref.
+- [sparkles-core] Add soft flushing threshold for thread local storage.
 - [encoder format] Add ticks_per_sec packet type
 - [sparkles-parser] Add small offset when several events recorded with the same timestamp
+
 
 ## [0.1.3] - 2024-09-23
 - [sparkles-core] New: Add configuration structures: `SparklesConfig` and `LocalStorageConfig`